+++ assets/files/localassets.go
@@ -51,9 +51,6 @@ func (la *LocalAssetBrowser) Browse(ctx context.Context) chan *assets.LocalAsset
 					return nil
 				}
 				if d.IsDir() {
-					if name != "." {
-						la.addAlbum(name)
-					}
 					return nil
 				}
 
@@ -72,10 +69,6 @@ func (la *LocalAssetBrowser) Browse(ctx context.Context) chan *assets.LocalAsset
 					f.Err = err
 				} else {
 					f.FileSize = int(s.Size())
-					dir := path.Dir(name)
-					if dir != "." {
-						f.Albums = []string{path.Base(dir)}
-					}
 					if f.DateTaken.IsZero() {
 						err = la.ReadMetadataFromFile(&f)
 						_ = err
@@ -131,59 +124,3 @@ func (la *LocalAssetBrowser) ReadMetadataFromFile(a *assets.LocalAssetFile) erro
 	}
 	return err
 }
-
-/*
-func (fsys LocalAssetBrowser) BrowseAlbums(ctx context.Context) error {
-	fsys.albums = map[string]string{}
-	err := fs.WalkDir(fsys, ".",
-		func(name string, d fs.DirEntry, err error) error {
-			if err != nil {
-				return err
-			}
-
-			// Check if the context has been cancelled
-			select {
-			case <-ctx.Done():
-				// If the context has been cancelled, return immediately
-				return ctx.Err()
-			default:
-			}
-			if name != "." && d.IsDir() {
-				fsys.albums[name] = fsys.albums[path.Base(name)]
-				return nil
-			}
-			return nil
-		})
-
-	return err
-}
-*/
-
-/*
-func ReadLocalAsset(fsys fs.FS, name string) (*assets.LocalAssetFile, error) {
-	ext := strings.ToLower(path.Ext(name))
-	switch ext {
-	case ".jpg", "jpeg", ".png", ".mp4", ".heic", ".mov", ".gif":
-		s, err := fs.Stat(fsys, name)
-		if err != nil {
-			return nil, fmt.Errorf("can't read asset: %w", err)
-		}
-		f := assets.LocalAssetFile{
-			FSys:     fsys,
-			FileName: name,
-			Title:    name,
-			size:     int(s.Size()),
-		}
-		_, err = fs.Stat(fsys, name+".xmp")
-		if err == nil {
-			f.SideCar = &metadata.SideCar{
-				FileName: name + ".xmp",
-				OnFSsys:  true,
-			}
-		}
-		return &f, nil
-	default:
-		return nil, fmt.Errorf("%q not supported: %q", ext, name)
-	}
-}
-*/