+++ assets/gp/testgp_test.go
@@ -168,9 +168,9 @@ func TestAlbums(t *testing.T) {
 			for a := range b.Browse(ctx) {
 				if len(a.Albums) > 0 {
 					for _, al := range a.Albums {
-						l := albums[al]
+						l := albums[al.Name]
 						l = append(l, fileResult{name: path.Base(a.FileName), size: a.FileSize, title: a.Title})
-						albums[al] = l
+						albums[al.Name] = l
 					}
 				}
 			}