+++ helpers/fshelper/multifs.go
@@ -2,6 +2,7 @@ package fshelper
 
 import (
 	"archive/zip"
+	"errors"
 	"io/fs"
 	"os"
 	"path/filepath"
@@ -19,16 +20,26 @@ func OpenMultiFile(names ...string) (fs.FS, error) {
 			return nil, err
 		}
 
-		switch {
-		case !s.IsDir() && strings.ToLower(filepath.Ext(s.Name())) == ".zip":
-			fsys, err := zip.OpenReader(p)
-			if err != nil {
-				return nil, err
-			}
-			fss = append(fss, fsys)
-		default:
+		if s.IsDir() {
 			fsys := DirRemoveFS(p)
 			fss = append(fss, fsys)
+		} else {
+			switch strings.ToLower(filepath.Ext(s.Name())) {
+			case ".zip":
+				fsys, err := zip.OpenReader(p)
+				if err != nil {
+					return nil, err
+				}
+				fss = append(fss, fsys)
+			case ".tgz":
+				return nil, errors.New("can't read directly .tgz files. Decompress it in a folder, an process the folder")
+			default:
+				fsys, err := newSingleFileFS(p)
+				if err != nil {
+					return nil, err
+				}
+				fss = append(fss, fsys)
+			}
 		}
 	}
 	return merged_fs.MergeMultiple(fss...), nil