+++ docs/releases.md
@@ -1,12 +1,37 @@
 # Release notes 
 
-## Release 0.5.1
+## Release next
 
 ### Fix #50:  Duplicate detection fails when timezone of both images differs
 Imported duplicated images with same name, but different timezone wasn't seen as duplicates.
 The `-ignore-tz-errors=true` compares the time on date, and minute and ignores the hour of capture.
 
 
+
+## Release 0.6.0
+
+New options for Google Phots albums:
+`-keep-untitled-albums <bool>` Untitled albums are imported into `immich` with the name of the folder as title (default: FALSE).<br>
+`-use-album-folder-as-name <bool>` Use the folder's name instead of the album title (default: FALSE).<br>
+
+### More integration tests
+- For the command upload
+- For Date based on file names
+
+
+### Fix #53 import from folders always creates albums
+Now albums are created only when requested.
+
+### Fix #48: Import from google takeout duplicates albums with special characters
+By default, Album are named after their title in JSON file, Special characters are allowed.
+The option `-use-album-folder-as-name=TRUE` names albums after the folder name instead of their title
+
+### Fix #42 [google photos] Lots of "No title" albums created with 1 file each  
+By default untitled albums are not created.
+Use the option `-keep-untitled-albums=TRUE` to keep them. 
+
+### Fix #51 Import a single file doesn't work
+It's now possible to import one file.
 
 ## Release 0.5.0